+++ FarmProject/dto/pressure-sensor/measurements/PressureMeasurementsFromSensorDto.cs
@@ -3,7 +3,7 @@
     public class PressureMeasurementsFromSensorDto
     {
         public required string IMEI { get; set; }
-        public required double PRR1 { get; set; }
-        public required double PRR2 { get; set; }
+        public required float PRR1 { get; set; }
+        public required float PRR2 { get; set; }
     }
 }
