+++ FarmProject/dto/pressure-sensor/settings/PressureSensorSettingsFromAdminClientDto.cs
@@ -2,12 +2,12 @@
 {
     public class PressureSensorSettingsFromAdminClientDto
     {
-        public required double DeviationSpanPositive { get; set; }
-        public required double DeviationSpanNegative { get; set; }
-        public required double Sensitivity { get; set; }
+        public required float DeviationSpanPositive { get; set; }
+        public required float DeviationSpanNegative { get; set; }
+        public required float Sensitivity { get; set; }
         public required bool AlarmActivated { get; set; }
         public required bool FirstSensorIsActive { get; set; }
         public required bool SecondSensorIsActive { get; set; }
-        public required double DataSendingSpan { get; set; }
+        public required float DataSendingSpan { get; set; }
     }
 }
