+++ FarmProject/dto/pressure-sensor/settings/PressureSensorSettingsToAdminClientDto.cs
@@ -2,13 +2,13 @@
 {
     public class PressureSensorSettingsToAdminClientDto
     {
-        public required double DeviationSpanPositive { get; set; }
-        public required double DeviationSpanNegative { get; set; }
-        public required double Sensitivity { get; set; }
+        public required float DeviationSpanPositive { get; set; }
+        public required float DeviationSpanNegative { get; set; }
+        public required float Sensitivity { get; set; }
         public required bool AlarmActivated { get; set; }
         public required bool FirstSensorIsActive { get; set; }
         public required bool SecondSensorIsActive { get; set; }
-        public required double DataSendingSpan { get; set; }
+        public required float DataSendingSpan { get; set; }
 
         public required string IMEI { get; set; }
     }
